Remove unused infantry and transporter logic.
