--- src/FieldWarning/Assets/Units/Component/Movement/InfantryMovementComponent.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 2971f9cb534a05a438ffccf13d718819
-timeCreated: 1463771045
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
