+++ src/FieldWarning/Assets/Units/Component/Movement/MoveMode.cs.meta
@@ -1,7 +1,5 @@
 fileFormatVersion: 2
-guid: e37eda7b78ad1354f8af131216bae12d
-timeCreated: 1533310221
-licenseType: Free
+guid: bda1f226beaa55944abb297141007ced
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
