--- src/FieldWarning/Assets/Units/Module.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: b789e236846edad43be42fa3cb1cd381
-folderAsset: yes
-timeCreated: 1533310221
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
