--- src/FieldWarning/Assets/Units/Module/MovementModule.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 63e5b6a4ef2542743a1a3309aa813073
-timeCreated: 1533310221
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
