--- src/FieldWarning/Assets/Units/Module/TransportableModule.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 2786a24155ea53d429a94e7e4b93328d
-timeCreated: 1533310221
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
