--- src/FieldWarning/Assets/Units/Module/TransporterModule.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 11f187ff685c7324db904d125c012128
-timeCreated: 1533310221
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
