--- src/FieldWarning/Assets/Units/TransporterBehaviour.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: c518b157c2b6adf4ba91b20a26e1b3b8
-timeCreated: 1463929851
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
