--- src/FieldWarning/Assets/Units/Waypoint.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 7862eb9f146d4dc498221ffedfcb937a
-folderAsset: yes
-timeCreated: 1533310221
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
