--- src/FieldWarning/Assets/Units/Waypoint/TransportableWaypoint.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 0458abe1870e032458e50d004f07fbfd
-timeCreated: 1533310221
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
