--- src/FieldWarning/Assets/Units/Waypoint/TransporterWaypoint.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 3b5a058a2a69f8d41bb754baae655b61
-timeCreated: 1533310221
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
