--- src/FieldWarning/Assets/Units/Waypoint/Waypoint.cs
@@ -1,31 +1,0 @@
-﻿/**
- * Copyright (c) 2017-present, PFW Contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
- * compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is
- * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See
- * the License for the specific language governing permissions and limitations under the License.
- */
-
-using PFW.Units;
-
-public abstract class Waypoint
-{
-    public bool interrupted;
-    public PlatoonBehaviour platoon;
-
-    public Waypoint(PlatoonBehaviour p)
-    {
-        platoon = p;
-    }
-
-    public abstract void ProcessWaypoint();
-
-    public abstract bool OrderComplete();
-
-    public abstract bool Interrupt();
-}
