--- src/FieldWarning/Assets/Units/Waypoint/Waypoint.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: f07a3abe07779ea43a09e719c015073a
-timeCreated: 1531653871
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
